requirements/_base.txt
envparse==0.2.0
kanboard==1.0.1
python-gitlab==0.20
